refactor: load DAO data
