+++ src/features/view-daos-in-network/components/DAOTableCard.module.scss
