+++ src/features/view-daos-in-network/components/index.ts
@@ -1,0 +1,1 @@
+export * from './DAOTable';
