+++ src/features/view-daos-in-network/index.ts
@@ -1,0 +1,1 @@
+export { DAOTable } from './components/DAOTable';
