+++ src/features/view-daos-in-network/lib/constants.ts
@@ -1,0 +1,5 @@
+export enum TABLE_KEYS {
+	TITLE = 'title',
+	ZNA = 'zna',
+	DAO = 'dao',
+}
