+++ src/features/view-daos-in-network/lib/types.ts
@@ -1,0 +1,8 @@
+import { zDAO } from '@zero-tech/zdao-sdk';
+import { TABLE_KEYS } from './constants';
+
+export type DAOTableDataItem = {
+	[TABLE_KEYS.TITLE]: string;
+	[TABLE_KEYS.ZNA]: string;
+	[TABLE_KEYS.DAO]: zDAO;
+};
