--- src/features/view-daos/DaosTable/index.ts
@@ -1,1 +1,0 @@
-export * from './DaosTable';
