--- src/features/view-daos/DaosTable/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './useDaosTableItemData';
