--- src/features/view-daos/index.ts
@@ -1,1 +1,0 @@
-export * from './DaosTable';
