+++ src/lib/hooks/queries/useAllDaos.ts
@@ -2,7 +2,7 @@ import { useQuery } from 'react-query';
 import { useAllZnas, useWeb3, useZdaoSdk } from '../';
 
 /**
- * Returns all DAOs on the current chain
+ * Returns all daos on the current chain
  */
 export const useAllDaos = () => {
 	const sdk = useZdaoSdk();
@@ -13,14 +13,13 @@ export const useAllDaos = () => {
 	const query = useQuery(
 		['daos', { chainId }],
 		async () => {
-			console.log('getting zDAO for ', znas);
 			return await Promise.all(znas.map((zna) => sdk.getZDAOByZNA(zna)));
 		},
 		{
 			retry: false,
 			refetchOnMount: false,
 			refetchOnWindowFocus: false,
-			enabled: Boolean(znas) && Boolean(znas.length),
+			enabled: Boolean(znas?.length),
 		},
 	);
