--- src/pages/DAOs/index.ts
@@ -1,1 +1,0 @@
-export * from './BrowseDAOs';
