+++ src/pages/daos/components/Page.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/breakpoints';
+@import '../../../styles/breakpoints';
 
 .Stats {
 	display: flex;
