+++ src/pages/daos/index.ts
@@ -1,0 +1,1 @@
+export { DAOsPage } from './components';
