+++ src/pages/daos/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './DaosTotalProvider';
+export * from './useDaosTotal';
